reorganize into input and output models and move extensions into classes 1-1 with the type they extend
