+++ tsp-output/@azure-tools/typespec-csharp/azoai/AzureOpenAI/Models/Input/AzureCreateChatCompletionRequest.Serialization.cs
