+++ tsp-output/@azure-tools/typespec-csharp/azoai/AzureOpenAI/Models/Input/AzureSearchChatExtensionConfiguration.Serialization.cs
