+++ tsp-output/@azure-tools/typespec-csharp/azoai/AzureOpenAI/Models/Input/AzureSearchChatExtensionParameters.Serialization.cs
