+++ tsp-output/@azure-tools/typespec-csharp/azoai/AzureOpenAI/Models/Input/UnknownAzureChatExtensionConfiguration.Serialization.cs
