+++ tsp-output/@azure-tools/typespec-csharp/azoai/AzureOpenAI/Models/Output/AzureChatCompletionResponseMessage.Serialization.cs
