+++ tsp-output/@azure-tools/typespec-csharp/azoai/AzureOpenAI/Models/Output/AzureCreateChatCompletionResponse.Serialization.cs
