+++ tsp-output/@azure-tools/typespec-csharp/azoai/AzureOpenAI/Models/Output/AzureCreateChatCompletionResponseChoice.Serialization.cs
